Now gets likes and dislikes of a post. Began work on getting parents
